Record data with multiple seeds and seed prefixes
The results seem dependent on seed, so I'm going to run the tests
for multiple seed prefixes and look at how much the list membership
changes. If the list isn't stable enough, I'll expand the number
of tests that are averaged. If the list is stable, I'll use it to
narrow down the important variables and exhaustively iterate to
identify the optimum configuration.